I did a heccin' ton of stuff, I'm not sure what
I made it not stop when encountering an error, added warnings, added macros, and when the user uses a variable that isn't defined, it will give a list of variables that are already defined and are similar to the undefined variable, ordered by similarity. I'm sure I did some other things, as well.